requirements.txt
praw==7.7.1
requests==2.32.3
pytz==2024.1
python-dateutil==2.9.0.post0
